+++ src/main/java/com/ward_n6/service/CatService.java
@@ -3,11 +3,10 @@ package com.ward_n6.service;
 import com.ward_n6.entity.pets.Cat_2;
 import com.ward_n6.enums.PetsSex;
 import com.ward_n6.repository.CatsCrud;
-
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -15,6 +14,10 @@ public class CatService {
 
     @Resource
     private CatsCrud catsCrud;
+
+    public CatService(CatsCrud catsCrud) {
+        this.catsCrud = catsCrud;
+    }
 
 
     public void addCat(PetsSex petsSex, Cat_2 cat_2) {
@@ -27,23 +30,15 @@ public class CatService {
         return byId.orElse(null);
     }
 
-    public String deleteCat(long id) {
-        try {
-            Cat_2 cat = findCat(id);
-            catsCrud.deleteById(id);
-            return cat.getPetName() + " удалена";
-        } catch (EmptyResultDataAccessException e) {
-            return "Такая кошка была удалена ранее или вы ошиблись с поиском.";
-        }
+    public Cat_2 deleteCat(long id) {
+        Cat_2 cat = findCat(id);
+        if (cat != null) catsCrud.deleteById(id);
+        return cat;
     }
 
 
-    public String allCats() {
-        StringBuilder stringBuilder = new StringBuilder();
-        for (Cat_2 cat2 : catsCrud.findAll()) {
-            stringBuilder.append("\n").append(cat2);
-        }
-        return stringBuilder.toString();
+    public List<Cat_2> allCats() {
+        return catsCrud.findAll();
     }
 
     public Cat_2 change(long id, Cat_2 cat_2, PetsSex petsSex) {
